secret-sync: extract secretSyncer into its own package
This commit extracts the secretSyncer into its own package.
